change healthRegen to int

+++ Assets/Scripts/Player/PlayerStats.cs
@@ -5,7 +5,7 @@ public class PlayerStats : MonoBehaviour
     [Header("Stats")]
     public int health = 3;
     public int maxHealth = 3;
-    public float healthRegen;
+    public int healthRegen;
     public float movementSpeed = 5f;
 
     // Start is called once before the first execution of Update after the MonoBehaviour is created
